chore(v0.9.2-1): retro-gate audit GHA + cargo-test isolation + release-deps watchlist
Three small P2 carryovers from the v0.9.1 release retro, bundled into one
PR per the v0.9.2-1 milestone scope.

#1234 — Pipeline-bypass CI check (ongoing). New scheduled GHA
  `.github/workflows/retro-gate-audit.yml` runs daily at 13:00 UTC and
  calls `scripts/audit-pipeline-bypass.py --limit 50` (the audit script
  shipped in PR #1229). Flagged PRs surface as workflow annotations in
  the Actions UI; no auto-issue creation to avoid spam during steady-state
  operation. Manual `workflow_dispatch` trigger included for ad-hoc
  audits and smoke testing. Part 2 of #1212.

#1235 — Isolated cargo-test target for filter_registry::tests.
  Moved the two filter_registry hot-path short-circuit tests from the
  in-module `#[cfg(test)]` block into
  `crates/djust_templates/tests/test_filter_registry_isolated.rs`. Cargo
  runs each integration-test file in its own process binary, so the
  process-global `ANY_CUSTOM_FILTERS_REGISTERED` AtomicBool starts as
  `false` for every run. The `OnceLock` workaround that gated the
  in-module test on whether a prior test had already registered a
  filter is no longer needed — the assertion is now ordering-independent
  and meaningful. The in-module test block is replaced with a comment
  pointing at the new file. Carryover from #1180 item 4.

#1236 — Watch-list for release-workflow-touching dep bumps. New
  `.github/workflows/check-release-workflow-deps.yml` runs on PRs
  modifying release-critical workflow files (release.yml, publish.yml,
  release-drafter.yml, pre-release-security-audit.yml). The workflow
  fails unless the PR carries the `release-workflow-reviewed` label,
  forcing explicit human risk-review before merge. Triggered by PR
  #1233 (action-gh-release v2 → v3) landing in the same window as the
  v0.9.1 cut — if v3 had a regression, the cut would have surfaced it
  first. The `release-workflow-reviewed` label is created in the repo
  as part of this change (gh label create).

Closes #1234.
Closes #1235.
Closes #1236.
